+++ src/GameClient.ts
@@ -1,9 +1,5 @@
 import { RunnerV1, RunnerV1Game, RunnerV2, RunnerV2Game, RunnerV3, RunnerV3Game } from "@akashic/headless-driver";
 import type { RunnerAdvanceConditionFunc, RunnerRenderingMode } from "@akashic/headless-driver";
-import type { ResourceFactory as ResourceFactoryV1 } from "@akashic/headless-driver/lib/runner/v1/platform/ResourceFactory";
-import type { ResourceFactory as ResourceFactoryV2 } from "@akashic/headless-driver/lib/runner/v2/platform/ResourceFactory";
-import type { NodeCanvasResourceFactory as ResourceFactoryV3_NodeCanvas } from "@akashic/headless-driver/lib/runner/v3/platform/NodeCanvasResourceFactory";
-import type { NullResourceFactory as ResourceFactoryV3_Null } from "@akashic/headless-driver/lib/runner/v3/platform/NullResourceFactory";
 import type { Canvas } from "canvas";
 import * as uuid from "uuid";
 import type { EngineVersions } from "./types";
@@ -63,29 +59,6 @@ export class GameClient<EngineVersion extends keyof EngineVersions = keyof Engin
 		this.type = type;
 		this.renderingMode = renderingMode;
 		this.g = runner.g;
-	}
-
-	private resourceFactoryClassV1: typeof ResourceFactoryV1;
-	private resourceFactoryClassV2: typeof ResourceFactoryV2;
-	private resourceFactoryClassV3: typeof ResourceFactoryV3_NodeCanvas | typeof ResourceFactoryV3_Null;
-
-	/**
-	 * 初期化処理
-	 */
-	async initialize(): Promise<void> {
-		// canvas を利用しない場合でも静的 import だと NodeCanvasResourceFactory で canvas を参照するので動的 import とする
-		const resfV1 = await import("@akashic/headless-driver/lib/runner/v1/platform/ResourceFactory");
-		const resfV2 = await import("@akashic/headless-driver/lib/runner/v2/platform/ResourceFactory");
-		this.resourceFactoryClassV1 = resfV1.ResourceFactory;
-		this.resourceFactoryClassV2 = resfV2.ResourceFactory;
-
-		if (this.renderingMode === "canvas") {
-			const nodeCanvasResf = await import("@akashic/headless-driver/lib/runner/v3/platform/NodeCanvasResourceFactory");
-			this.resourceFactoryClassV3 = nodeCanvasResf.NodeCanvasResourceFactory;
-		} else {
-			const nullResf = await import("@akashic/headless-driver/lib/runner/v3/platform/NullResourceFactory");
-			this.resourceFactoryClassV3 = nullResf.NullResourceFactory;
-		}
 	}
 
 	/**
@@ -175,12 +148,13 @@ export class GameClient<EngineVersion extends keyof EngineVersions = keyof Engin
 		const resFac = this.runner.platform.getResourceFactory();
 		const id = param.id ?? uuid.v4();
 		const path = param.path ?? uuid.v4();
+		const version = this.runner.engineVersion;
 
-		if (resFac instanceof this.resourceFactoryClassV1) {
+		if (version === "1") {
 			return resFac.createImageAsset(id, path, param.width, param.height);
-		} else if (resFac instanceof this.resourceFactoryClassV2) {
+		} else if (version === "2") {
 			return resFac.createImageAsset(id, path, param.width, param.height);
-		} else if (resFac instanceof this.resourceFactoryClassV3) {
+		} else if (version === "3") {
 			return resFac.createImageAsset(id, path, param.width, param.height);
 		}
 
@@ -197,15 +171,16 @@ export class GameClient<EngineVersion extends keyof EngineVersions = keyof Engin
 		const path = param.path ?? uuid.v4();
 		const loop = !!param.loop;
 
-		if (resFac instanceof this.resourceFactoryClassV1) {
+		const version = this.runner.engineVersion;
+		if (version === "1") {
 			const game = this.game as RunnerV1Game;
 			const system = param.systemId ? game._audioSystemManager[param.systemId] : game._audioSystemManager[game.defaultAudioSystemId];
 			return resFac.createAudioAsset(id, path, param.duration, system, loop, param.hint);
-		} else if (resFac instanceof this.resourceFactoryClassV2) {
+		} else if (version === "2") {
 			const game = this.game as RunnerV2Game;
 			const system = param.systemId ? game._audioSystemManager[param.systemId] : game._audioSystemManager[game.defaultAudioSystemId];
 			return resFac.createAudioAsset(id, path, param.duration, system, loop, param.hint);
-		} else if (resFac instanceof this.resourceFactoryClassV3) {
+		} else if (version === "3") {
 			const game = this.game as RunnerV3Game;
 			const system = param.systemId ? game.audio[param.systemId] : game.audio[game.defaultAudioSystemId];
 			return resFac.createAudioAsset(id, path, param.duration, system, loop, param.hint);